scaling

+++ src/tests_metric.c
@@ -6,11 +6,11 @@ void test_metric_equal_graphs(int* passed, int* failed) {
     graph_generate(g1, 7, 1, 0.8f, 1);
     matrix* g2 = matrix_clone(g1);
 
-    int d = distance(g1, g2);
+    float d = distance(g1, g2);
 
     graph_print(g1, "Graph 1");
     graph_print(g2, "Graph 2");
-    printf("\nDistance: %d", d);
+    printf("\nDistance: %f", d);
 
     if (d == 0) {
         print_test_pass(__func__);
@@ -35,9 +35,9 @@ void test_metric_permuted_graphs(int* passed, int* failed) {
     graph_print(g1, "Graph 1");
     graph_print(g2, "Graph 2");
 
-    int d = distance(g1, g2);
+    float d = distance(g1, g2);
 
-    printf("\nDistance: %d", d);
+    printf("\nDistance: %f", d);
 
     if (d == 0) {
         print_test_pass(__func__);
@@ -60,11 +60,11 @@ void test_metric_different_graphs(int* passed, int* failed) {
 
     graph_add_noise(g2, 0.5f, 1, 0.0);
 
-    int d = distance(g1, g2);
+    float d = distance(g1, g2);
 
     graph_print(g1, "Graph 1");
     graph_print(g2, "Graph 2");
-    printf("\nDistance: %d", d);
+    printf("\nDistance: %f", d);
 
     if (d != 0) {
         print_test_pass(__func__);
@@ -97,12 +97,12 @@ void tests_metric(int* passed, int* failed) {
 void test_metric_from_args(matrix* g1, matrix* g2, int* passed, int* failed) {
     printf("\n\nRunning metrics test...\n");
 
-    int d = distance(g1, g2);
+    float d = distance(g1, g2);
 
     graph_print(g1, "Graph 1");
     graph_print(g2, "Graph 2");
 
-    printf("\nDistance: %d\n", d);
+    printf("\nDistance: %f\n", d);
 
     (*passed)++;
     print_test_pass(__func__);
